Complete rewrite of the Elantech clickpads v4 packet parser in order to support "click-and-drag".
This implements a simple filter to keep track of a moving finger on the
clickpad.  For the moment wsmouse(4) and wscons(4) are still unaware of
multitouch events and pms(4) has to do the translation, but this is a
small steps in this direction.

All the good work has been done by Ulf Brosziewski.