Call if_update() only when necessary for IPv6 address changes.
For EIGRPv6, an interface only needs an link-local address in order
to be enabled in the routing daemon. So adding or removing global
unicast addresses, for example, will never trigger a state change in
the interface.

Additionally, move all address handling code to if_addr_new() and
if_addr_del() to improve readability.