I2C driver for the Maxim DS1307 Real Time Clock chip.
The chip has been found on a Sun Fire V210 as a second RTC hooked up
to the SMU.  Setting the time there has the advantage that it remains
after a boot since the internal V210 RTC will be synced by the DS1307
at boot time.

Not linked to sparc64 yet since it will overwrite the internal RTC
handler.

ok deraadt