Remove a feature to re-use existing early bootstrap mappings. This allowed you to pass a virtual address, which you received from an early bootstrap mapping, as physical address to bus_space_map(9).
It breaks bus_space_map(9) for peripherals that are after 0xC0000000,
as it assumes that everything after that address cannot be a real
peripheral.  But that's wrong.  It does not make sense to pass a
virtual address to bus_space_map(9) anyway, so just get rid of this
whole "feature".

ok kettenis@