Dynamically attach ampintc(4) and make it register itself as an interrupt controller.  Switch all i.MX6 devices over to the new FDT-aware interrupt establish API and enable imxgpc(4).
This may break other platforms that use ampintc(4) as their interrupt
controller.  We will fix this as soon as possible by switching them over
to the new interrupt establish API as well.

This also removes the hack in fec(4) to avoid the gpio-based interrupt
workaround.  This commit is an essential step on the road to fix that
issue properly.

ok patrick@, jsg@